Add safe call to products and destroy inactive categories

+++ app/services/integration/tools/categories_cleanup.rb
@@ -5,21 +5,23 @@ module Integration
     class CategoriesCleanup
       def initialize(zecat_country)
         @zecat_country = zecat_country
-        @zecat_categories = CountrySelection::zecat_class_name(zecat_country)::Families.categories
+        @zecat_categories = CountrySelection.zecat_class_name(zecat_country)::Families.categories
       end
 
       def iterate_categories_and_destroy
         Category.where(country: @zecat_country).each_slice(10) do |group|
           group.each do |local_category|
-            zecat_category = CountrySelection::zecat_class_name(@zecat_country)::Families.category_by_id(@zecat_categories, local_category.zecat_id.to_s)
-            woocommerce_category = CountrySelection::woocommerce_class_name(@zecat_country).category_by_id(local_category.woocommerce_api_id)
+            zecat_category = CountrySelection.zecat_class_name(@zecat_country)::Families.category_by_id(
+              @zecat_categories, local_category.zecat_id.to_s
+            )
+            woocommerce_category = CountrySelection.woocommerce_class_name(@zecat_country).category_by_id(local_category.woocommerce_api_id)
+
+            local_category.destroy! if (zecat_category&.dig('show') == false) && !zecat_category.nil?
 
             next if !zecat_category.nil? && woocommerce_category.present?
 
             if !zecat_category.nil? && !woocommerce_category.present?
               local_category.update(last_sync: nil)
-            elsif zecat_category&.dig('show') == false
-              local_category.destroy!
             elsif zecat_category.nil?
               local_category.destroy!
             end
